Extract BoundVersion to separate file
