+++ Sources/PackageGraph/CMakeLists.txt
@@ -7,6 +7,7 @@
 # See http://swift.org/CONTRIBUTORS.txt for Swift project authors
 
 add_library(PackageGraph
+  BoundVersion.swift
   CheckoutState.swift
   DependencyResolver.swift
   LocalPackageContainer.swift
